requirements.txt
numpy==1.26.3
onnxruntime==1.16.3
gradio==5.9.1
Pillow==10.1.0
huggingface-hub>=0.25.1
pydantic>=2.0